chore: convert command registry to a class
